requirements.txt
flask==3.0.0
werkzeug==3.0.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
nltk==3.8.1
matplotlib==3.8.2
seaborn==0.13.0
scipy==1.11.4
python-dotenv==1.0.0
gunicorn==21.2.0
